requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
defusedxml==0.5.0         # via python3-openid
django-allauth==0.35.0
django-appconf==1.0.2     # via django-compressor
django-braces==1.12.0
django-compressor==2.2
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-extensions==2.0.0
django-tables2==1.19.0
django==2.0.2
idna==2.6                 # via requests
oauthlib==2.0.6           # via requests-oauthlib
python3-openid==3.1.0     # via django-allauth
pytz==2018.3              # via django
rcssmin==1.0.6            # via django-compressor
requests-oauthlib==0.8.0  # via django-allauth
requests==2.18.4          # via django-allauth, requests-oauthlib
rjsmin==1.0.12            # via django-compressor
six==1.11.0               # via django-extensions
sqlparse==0.2.4           # via django-debug-toolbar
typing==3.6.4             # via django-extensions
urllib3==1.22             # via requests
whitenoise==3.3.1
